requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
charset-normalizer==2.0.9
discord==1.7.3
discord.py==1.7.3
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pendulum==2.1.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytzdata==2020.1
six==1.16.0
typing-extensions==4.0.1
yarl==1.7.2
